move Commons-IO FilenameUtils to own class
